+++ composer.json
@@ -27,7 +27,7 @@
     "require-dev": {
         "behat/behat": "~2.4",
         "monolog/monolog": "~1.3",
-        "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "~4.0"
     },
     "suggest": {
         "monolog/monolog": "The recommended logging library to use with Propel."
