+++ src/Propel/Runtime/Connection/ConnectionWrapper.php
@@ -32,6 +32,8 @@ use Psr\Log\LoggerAwareInterface;
  */
 class ConnectionWrapper implements ConnectionInterface, LoggerAwareInterface
 {
+    use TransactionTrait;
+
     /**
      * Attribute to use to set whether to cache prepared statements.
      */
@@ -302,23 +304,6 @@ class ConnectionWrapper implements ConnectionInterface, LoggerAwareInterface
     public function inTransaction()
     {
         return $this->connection->inTransaction();
-    }
-
-    /**
-     * Executes the given callable within a transaction.
-     * This helper method takes care to commit or rollback the transaction.
-     *
-     * In case you want the transaction to rollback just throw an Exception of any type.
-     *
-     * @param callable $callable A callable to be wrapped in a transaction
-     *
-     * @return mixed Returns the result of the callable.
-     *
-     * @throws \Exception Re-throws a possible <code>Exception</code> triggered by the callable.
-     */
-    public function transaction(callable $callable)
-    {
-        return $this->connection->transaction($callable);
     }
 
     /**
